db status: handle all(?) branch cases
