+++ backend/geonature/core/command/__init__.py
@@ -1,2 +1,3 @@
 from .main import main
+from .alembic import db_cli
 import geonature.core.command.create_gn_module
