+++ backend/geonature/core/command/main.py
@@ -4,26 +4,13 @@
 
 import logging
 from os import environ
-from collections import ChainMap, deque
-from io import StringIO
+from collections import ChainMap
 
 import toml
 import click
 from flask.cli import run_command
-import flask_migrate
-import alembic
-from alembic.migration import MigrationContext
-from alembic.context import EnvironmentContext
-from alembic.script import ScriptDirectory
-from flask_migrate.cli import db as db_cli
-from flask.cli import with_appcontext
 
-from geonature.utils.env import (
-    db,
-    migrate,
-    DEFAULT_CONFIG_FILE,
-    GEONATURE_VERSION,
-)
+from geonature.utils.env import GEONATURE_VERSION
 from geonature.utils.command import (
     start_geonature_front,
     build_geonature_front,
@@ -191,138 +178,3 @@ def default_config():
     frontend_defaults = GnGeneralSchemaConf().load({}, partial=required_fields)
     defaults = ChainMap(backend_defaults, frontend_defaults)
     print(toml.dumps(defaults))
-
-
-@db_cli.command()
-@click.option('-d', '--directory', default=None,
-              help=('Migration script directory (default is "migrations")'))
-@click.option('--sql', is_flag=True,
-              help=('Don\'t emit SQL to database - dump to standard output '
-                    'instead'))
-@click.option('--tag', default=None,
-              help=('Arbitrary "tag" name - can be used by custom env.py '
-                    'scripts'))
-@click.option('-x', '--x-arg', multiple=True,
-              help='Additional arguments consumed by custom env.py scripts')
-@with_appcontext
-def autoupgrade(directory, sql, tag, x_arg):
-    config = migrate.get_config(directory, x_arg)
-    script = ScriptDirectory.from_config(config)
-    heads = set(script.get_heads())
-    migration_context = MigrationContext.configure(db.session.connection())
-    current_heads = migration_context.get_current_heads()
-    # get_current_heads does not return implicit revision through dependecies, get_all_current does
-    current_heads = set(map(lambda rev: rev.revision, script.get_all_current(current_heads)))
-    for head in current_heads - heads:
-        revision = head + '@head'
-        flask_migrate.upgrade(directory, revision, sql, tag, x_arg)
-
-
-@db_cli.command()
-@click.option('-d', '--directory', default=None,
-              help=('Migration script directory (default is "migrations")'))
-@click.option('-x', '--x-arg', multiple=True,
-              help='Additional arguments consumed by custom env.py scripts')
-@with_appcontext
-def status(directory, x_arg):
-    """Show all revisions sorted by branches."""
-    config = migrate.get_config(directory, x_arg)
-    script = ScriptDirectory.from_config(config)
-    migration_context = MigrationContext.configure(db.session.connection())
-
-    current_heads = migration_context.get_current_heads()
-    applied_rev = set(script.iterate_revisions(current_heads, 'base'))
-
-    bases = [ script.get_revision(base) for base in script.get_bases() ]
-    heads = [ script.get_revision(head) for head in script.get_heads() ]
-
-    outdated = False
-    for branch_base in sorted(bases, key=lambda rev: next(iter(rev.branch_labels))):
-        output = StringIO()
-        branch, = branch_base.branch_labels
-        levels = { branch_base: 0 }
-        branch_outdated = False
-        seen = set()
-        todo = deque()
-        todo.append(branch_base)
-        while todo:
-            rev = todo.pop()
-            seen.add(rev)
-            i = levels[rev]
-            if rev.is_branch_point:
-                for j, nextrev in enumerate(rev.nextrev):
-                    nextrev = script.get_revision(nextrev)
-                    if j == 0:
-                        symbol = ' ' * i + '┣'
-                        _i = i
-                        levels[nextrev] = _i
-                    else:
-                        __i = max(levels.values()) + 1
-                        symbol += '━' * (__i-_i-1)
-                        if j == len(rev.nextrev)-1:
-                            symbol += '┓'
-                        else:
-                            symbol += '┳'
-                        _i = __i
-                        levels[nextrev] = _i
-                    todo.append(nextrev)
-            elif rev.is_merge_point:
-                down_revision = { script.get_revision(r) for r in rev.down_revision }
-                if not down_revision.issubset(seen):
-                    continue
-                for j, downrev in enumerate(sorted(down_revision, key=lambda rev: levels[rev])):
-                    if j == 0:
-                        _i = levels[downrev]
-                        symbol = ' ' * _i
-                        if rev in heads:
-                            symbol += '┗'
-                        else:
-                            symbol += '┣'
-                    else:
-                        __i = levels[downrev]
-                        symbol += '━' * (__i-_i-1)
-                        _i = __i
-                        if j == len(rev.down_revision)-1:
-                            symbol += '┛'
-                        else:
-                            symbol += '┻'
-            else:
-                symbol = ' ' * i
-                if rev in heads and rev in bases:
-                    symbol += '─'
-                elif rev in heads:
-                    symbol += '┸'
-                elif rev in bases:
-                    symbol += '┰'
-                else:
-                    symbol += '┃'
-
-            if not rev.is_branch_point and rev.nextrev:
-                nextrev, = rev.nextrev
-                nextrev = script.get_revision(nextrev)
-                levels[nextrev] = i
-                todo.append(nextrev)
-
-            check = 'x' if rev in applied_rev else ' '
-            if branch_base in applied_rev and rev in applied_rev:
-                fg = 'white'
-            elif branch_base in applied_rev:
-                outdated = True
-                branch_outdated = True
-                fg = 'red'
-            else:
-                fg = None
-            print(click.style(f"  [{check}] {symbol} {rev.revision} {rev.doc}", fg=fg), file=output)
-
-        if branch_base in applied_rev:
-            fg = 'white'
-            mark = ' '
-            mark += click.style('×', fg='red') if branch_outdated else click.style('✓', fg='green')
-        else:
-            fg = None
-            mark = ''
-        click.echo(click.style(f"[{branch}", bold=True, fg=fg) + mark + click.style("]", bold=True, fg=fg))
-        click.echo(output.getvalue(), nl=False)
-
-    if outdated:
-        click.secho("Some branches are outdated, you can upgrade with: geonature db autoupgrade", fg="red")
